# Ammonia models are now faster!