So I've been profiling the ammonia model performance in `pyspeckit`, and, long story short, found out that both models and spectral fits can be improved to yield a large speed boost.

Here are the two areas I've addressed so far:

First of all, whenever `ammonia` function gets called, it tries to convert `xarr` into `GHz`. Somehow, this isn't optimized for speed in `astropy`, so in the end it ends up eating up a significant chunck of CPU power. This is fixed in 01142ec, and the `%timeit` results are given below in rows 2 and 3.

Now on to the largest performance boost. There's a `line_names` argument that gets passed to `ammonia` function, that, in turn, gets inherited by `_ammonia_spectrum`, specifying the inversion levels to calculate over. There are a few issues with this:

- [feature, not a bug] Because `line_names` defaults to all lines in (1,1)-(8,8) range, the default model that is registered under `Registry.multifitters['ammonia']` is slower than it should be for cubes that have, for example, only (1,1) and (2,2) lines. This can, of course, be solved by registering a new fitter with `function=ammonia_model(line_names=['oneone', 'twotwo'])` passed to the `add_fitter` method, but... I suspect that not many users figured it out. In fact, when I look over to the some data analysis pipelines for NH3 (e.g. GAS survey repo), they don't really use it. if you check the table below, the second and third columns show the time it takes to run `get_full_model` on the (1,1) and (2,2) ammonia spectrum from a typical GBT spectral setup. Clearly, the `ammonia_model` with matching 
`line_names` wins the race!

- However, selecting a proper `ammonia_model` had no effect on the fitting performance. That was, probably, a bug - the fitting function didn't pick up the model `kwargs` that were stored under `fitter.modelfunc_kwargs`. Therefore, even for `line_names=['oneone', 'twotwo']`, the fitter ran with all the inversion lines up to (8,8) regardless of the ammonia model registered, taking the defaults from the ammonia function. I set up the fitting kwargs to update from model ones in 508d946.

| `faster-ammonia` branch commit         | `get_full_model` for `line_names`=(1,1)-(2,2) | `get_full_model` for `line_names`=(1,1)-(8,8) | Fitting a dummy cube for `line_names`=(1,1)-(2,2) | Fitting a dummy cube for `line_names`=(1,1)-(8,8) |
|----------------------------------------|-----------------------------------------------|-----------------------------------------------|---------------------------------------------------|---------------------------------------------------|
| 71552fd (current master)      | 4.62 ms                                       | 9.76 ms                                       | 51.4 s                                            | 51.4 s                                            |
| 01142ec (condition call to as_unit)  | 3.33 ms                                       | 8.44 ms                                       | 40.8 s                                            | 40.8 s                                            |
| 508d946 (fitter inherits model kwargs) | -                                             | -                                             |  **26.7 s**   | 40.9 s                                            |


So, main takeaway is that fitting ammonia cubes can be twice as fast now in some cases!

@keflavich  I am still working on more improvements, so please don't merge just yet.