@keflavich 2017-03-21
Nice work!  I'll review when you're done.  These little speedups are a great help.

Yes, re-converting units every time is a bad thing.  So is computing the line profile for unobserved regions.