@vlas-sokolov 2017-03-22
All right, I wrote a few more improvements. I'm using [this script](https://gist.github.com/vlas-sokolov/af058f7e46a67c77fc796c614b5c859e) for timings and setting up models/cubes/spectra, in case anyone wants to experiment with it.

I haven't yet added the commit (40ad5ed) from `faster-mpfit' branch as it involves making changes into mpfit.py - is it okay to do so or is it an untouchable legacy code? Apart from that, #223 fixes some issues I had with lmfit, and together with changes on this PR there's a nice speedup achievable as well.

Both timings, for modified mpfit and for lmfit, are below:

| Branch name           | latest `master`   | `faster-ammonia` (this PR) | `faster-mpfit` with changes from this PR | `lmfit-properpars` (#223) with changes from this PR |
|-----------------------|----------|--------------------------|----------------------------------------|--------------------------------------------|
| Time of `fiteach` run | 51.6 sec | 27.9 sec                 | 20.9 sec                               | 14.9 sec (with `use_lmfit=True`)             |

Also, I wrote a version of `_ammonia_spectrum` function that uses Einstein summation instead of a for loop over the hyperfines (12963a9). It improves the running time of `get_full_model` and `fiteach(use_lmfit=True)` by ~5%, but for `fiteach(use_lmfit=False)` it degrades the performance by as much as a third. I'll try to figure out why, but for now it's probably better to leave it out of the PR.