@keflavich 2017-03-23
`mpfit.py` is touchable; pyspeckit contains a fork of that code (and it could probably do with a *lot* of optimizations).

If you make the einstein summation change, include some extra documentation with it (i.e., comments).