@vlas-sokolov 2017-03-23
Okay, I updated the PR with the mpfit tinkering - two conversions from `sum()` to `np.sum()`. It's possible that `mpfit.py`can be made even faster - I just fixed two lines that the profiler thought to be performance bottlenecks. Meanwhile, I'll try to see if I can find the cause for the strange `einsum` behavior...

@keflavich do you think it's worth investigating automatic selection of `line_names` in ammonia models?
 As I have shown above, proper `line_names` result in a major speed boost, but the problem is that, by default, a registered model doesn't know anything about the `xarr` instance it will act upon. However, checking for valid lines in `xarr` at every point a spectrum is calculated would - I'm guessing here - result in a much slower performance (astropy functions tend to be a bit sluggish).