@keflavich 2017-04-22
@vlas-sokolov should I merge this?  It looks fine; I had nothing to add in review.

Do you know how to run `line_profiler`?  That's the other way we could try to find places to squeeze out a little more performance.