@vlas-sokolov 2017-04-22
@keflavich I used `statprof` and `pycallgraph` for profiling, they both seem to do the same thing `line_profiler` does.

I think this can be merged. I haven't had the time to look into the `np.einsum` improvement, but I will leave that branch on GitHub for any potential future improvements.